DatabaseHandler for higher cohesion
Added a databasehandler inorder to increase cohesion
